refactor: Move CLI commands into their own files
